+++ src/pages/home/panel-boczny.css
@@ -17,7 +17,6 @@
     margin-top: 20px !important;
 }
 
-/* wyłącza podkreślenie do linków + nie nadpisuje oryginalnego koloru czcionki */
 a {
     text-decoration: none;
     color: inherit;
