+++ src/pages/home/searchUser.js
@@ -1,7 +1,8 @@
 // połączenie z bazą danych na Firebase
 import React, { useState } from 'react';
 import { getUserUid } from '../../firebase';
-import { Link, RouterProvider, useNavigate } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
+import SearchIcon from '@mui/icons-material/Search';
 
 function SearchUser()
 {
@@ -22,17 +23,20 @@ function SearchUser()
     };
 
     return(
-        <>
-        <form onSubmit={handleUserNameSubmit} autoComplete="off">
-                <input
-                    type="text"
-                    value={userName}
-                    onChange={handleUserNameChange}
-                    placeholder="Wyszukaj użytkownika..."
-                />
-                <button>Szukaj</button>
-        </form>
-        </>
+    <form className="search-form" onSubmit={handleUserNameSubmit} autoComplete="off">
+      <div className="search-container">
+        <input
+          className="search-input"
+          type="text"
+          value={userName}
+          onChange={handleUserNameChange}
+          placeholder="Wyszukaj użytkownika..."
+        />
+        <button className="search-button">
+          <SearchIcon /> 
+        </button>
+      </div>
+    </form>
     )
 }
